\name{SFT-package}
\alias{SFT-package}
\alias{SFT}
\docType{package}
\title{
\packageTitle{SFT}
}
\description{
\packageDescription{SFT}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SFT}
\packageIndices{SFT}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SFT}

Maintainer: \packageMaintainer{SFT}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
